requirements.txt
aniso8601==8.0.0
boto3==1.15.6
botocore==1.18.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.2
python-dateutil==2.8.1
python-lambda-local==0.1.12
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
